+++ manul/benches/empty_rounds.rs
@@ -7,9 +7,9 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use manul::{
     dev::{run_sync, BinaryFormat, TestSessionParams, TestSigner},
     protocol::{
-        Artifact, BoxedRound, CommunicationInfo, Deserializer, DirectMessage, EchoBroadcast, EntryPoint,
+        Artifact, BoxedFormat, BoxedRound, CommunicationInfo, DirectMessage, EchoBroadcast, EntryPoint,
         FinalizeOutcome, LocalError, MessageValidationError, NoProtocolErrors, NormalBroadcast, PartyId, Payload,
-        Protocol, ProtocolMessage, ProtocolMessagePart, ReceiveError, Round, RoundId, Serializer, TransitionInfo,
+        Protocol, ProtocolMessage, ProtocolMessagePart, ReceiveError, Round, RoundId, TransitionInfo,
     },
     signature::Keypair,
 };
@@ -24,7 +24,7 @@ impl<Id> Protocol<Id> for EmptyProtocol {
     type ProtocolError = NoProtocolErrors;
 
     fn verify_direct_message_is_invalid(
-        _deserializer: &Deserializer,
+        _format: &BoxedFormat,
         _round_id: &RoundId,
         _message: &DirectMessage,
     ) -> Result<(), MessageValidationError> {
@@ -32,7 +32,7 @@ impl<Id> Protocol<Id> for EmptyProtocol {
     }
 
     fn verify_echo_broadcast_is_invalid(
-        _deserializer: &Deserializer,
+        _format: &BoxedFormat,
         _round_id: &RoundId,
         _message: &EchoBroadcast,
     ) -> Result<(), MessageValidationError> {
@@ -40,7 +40,7 @@ impl<Id> Protocol<Id> for EmptyProtocol {
     }
 
     fn verify_normal_broadcast_is_invalid(
-        _deserializer: &Deserializer,
+        _format: &BoxedFormat,
         _round_id: &RoundId,
         _message: &NormalBroadcast,
     ) -> Result<(), MessageValidationError> {
@@ -80,7 +80,7 @@ impl<Id: PartyId> EntryPoint<Id> for Inputs<Id> {
 
     fn make_round(
         self,
-        _rng: &mut impl CryptoRngCore,
+        _rng: &mut dyn CryptoRngCore,
         _shared_randomness: &[u8],
         _id: &Id,
     ) -> Result<BoxedRound<Id, Self::Protocol>, LocalError> {
@@ -108,11 +108,11 @@ impl<Id: PartyId> Round<Id> for EmptyRound<Id> {
 
     fn make_echo_broadcast(
         &self,
-        _rng: &mut impl CryptoRngCore,
-        serializer: &Serializer,
+        _rng: &mut dyn CryptoRngCore,
+        format: &BoxedFormat,
     ) -> Result<EchoBroadcast, LocalError> {
         if self.inputs.echo {
-            EchoBroadcast::new(serializer, Round1EchoBroadcast)
+            EchoBroadcast::new(format, Round1EchoBroadcast)
         } else {
             Ok(EchoBroadcast::none())
         }
@@ -120,46 +120,42 @@ impl<Id: PartyId> Round<Id> for EmptyRound<Id> {
 
     fn make_direct_message(
         &self,
-        _rng: &mut impl CryptoRngCore,
-        serializer: &Serializer,
+        _rng: &mut dyn CryptoRngCore,
+        format: &BoxedFormat,
         _destination: &Id,
     ) -> Result<(DirectMessage, Option<Artifact>), LocalError> {
-        let dm = DirectMessage::new(serializer, Round1DirectMessage)?;
+        let dm = DirectMessage::new(format, Round1DirectMessage)?;
         let artifact = Artifact::new(Round1Artifact);
         Ok((dm, Some(artifact)))
     }
 
     fn receive_message(
         &self,
-        deserializer: &Deserializer,
+        format: &BoxedFormat,
         _from: &Id,
         message: ProtocolMessage,
     ) -> Result<Payload, ReceiveError<Id, Self::Protocol>> {
         if self.inputs.echo {
-            let _echo_broadcast = message
-                .echo_broadcast
-                .deserialize::<Round1EchoBroadcast>(deserializer)?;
+            let _echo_broadcast = message.echo_broadcast.deserialize::<Round1EchoBroadcast>(format)?;
         } else {
             message.echo_broadcast.assert_is_none()?;
         }
         message.normal_broadcast.assert_is_none()?;
-        let _direct_message = message
-            .direct_message
-            .deserialize::<Round1DirectMessage>(deserializer)?;
+        let _direct_message = message.direct_message.deserialize::<Round1DirectMessage>(format)?;
         Ok(Payload::new(Round1Payload))
     }
 
     fn finalize(
-        self,
-        _rng: &mut impl CryptoRngCore,
+        self: Box<Self>,
+        _rng: &mut dyn CryptoRngCore,
         payloads: BTreeMap<Id, Payload>,
         artifacts: BTreeMap<Id, Artifact>,
     ) -> Result<FinalizeOutcome<Id, Self::Protocol>, LocalError> {
         for payload in payloads.into_values() {
-            let _payload = payload.try_to_typed::<Round1Payload>()?;
+            let _payload = payload.downcast::<Round1Payload>()?;
         }
         for artifact in artifacts.into_values() {
-            let _artifact = artifact.try_to_typed::<Round1Artifact>()?;
+            let _artifact = artifact.downcast::<Round1Artifact>()?;
         }
 
         if self.round_counter == self.inputs.rounds_num {